More s/malloc(size) + init/calloc(1, size)/ replacements
This time in common/comments.c and common/downtime.c. Since we
really don't want to use explicit initialization on memory that
in all likelihood is already nul'ed out, as memory returned by
calloc() always is, and memory returned from malloc() that has
not been re-used by libc but comes straight from the kernel is,
this is a good thing.

Signed-off-by: Andreas Ericsson <ae@op5.se>

Author: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1495 5f96b256-904b-4d8d-8c98-d829582c6739