Fix some compiler warnings
They're mostly harmless and mostly relate to superfluous casts
while using the wrong type to begin with. Nothing was really
broken, except for possibly the embedded perl interpreter which
might have used an uninitialize cache pointer if caching and epn
was enabled. If it actually was a problem, it's fixed now though.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@1544 5f96b256-904b-4d8d-8c98-d829582c6739