Fix race on flexible downtime with invalid duration
it may happen that an external command sets a
flexible downtime with duration not set or zero.

a flexible downtime is triggered between start and
end time and needs a duration to be set and non zero.

without the fix, the scheduled downtime starts and then
tries to calculate the end time out of the set duration
for the downtime and scheduling a new event.
the time passed over there is in the past, and causes the
core to get into a loop.

to prevent this misbehavior, a duration of null or zero is
not allowed and triggered as error in base/commands.c not
setting any flexible downtime.

Signed-off-by: Michael Friedrich <michael.friedrich@univie.ac.at>
Signed-off-by: Andreas Ericsson <ae@op5.se>