Reduce locking of the checker singleton to avoid dead lock.
Two threads are running concurrently in Centreon Engine. The first
one is the main event loop. The second one comes from the Clib and
handle processes. Notably it calls the finished() method of the
checker as soon as a process terminates.

The first and the second thread were blocked because the first was
trying to launch a command (and locking the list of running
commands to add it) and the second was trying to process a
terminated command (and locking the list of running commands to
remove it).

The fix was to put the partial result of a terminated command in
a separate list that will be later processed by the event loop
(check reaping) which will also remove the process from the list
of running processes. This fixes #5466.