Show an error when objects use illegal object chars.
This bug was caused by the two competing global variables
illegal_object_chars. The first is the old C-style global
variable kept for compatibility and the other one is the field
from the global /config/ object introduced by the reload
mechanism. Compatibility globals are filled at the beginning of
the configuration application whereas the config object is filled
at the end, hence the problem. This fixes #5853.