feat(customvariables): Two variables added in centengine.cfg
By default all customvariables are sent to broker.
Now, in centengine, we can filter which custom variable to send.
For that purpose, we have two variables:
* enable_macros_filter that is a boolean
* macros_filter that is a string of variables separated by commas.

A new object custom_variable has been created and contains a flag
telling if the variable must be transfered to the central or not.

This change needs also a change in broker.