feat(simu): A new module is added to centreon-engine.
It is connected to engine as cbmod would be.
The only thing done by this module is to write on screen received events.

There's also a script provided check.sh that takes two arguments:

check.sh host 0
Here to tell that it is a host check that sends the host is UP.

check.sh host 1
Here to tell that it is a host check that sends the host is DOWN.

check.sh service 0
Here to tell that it is a service check that sends the service is OK.

check.sh service 1
Here to tell that it is a service check that sends the service is WARNING.

check.sh service 2
Here to tell that it is a service check that sends the service is CRITICAL.

Usual files of centreon can be seen here at the same place prefixed by /tmp/.